requirements.txt
pycryptodomex~=3.14
pynacl~=1.4
